Add lcms2-2.17-1
Little CMS intends to be an Open Source small-footprint color management engine, with special focus on accuracy and performance.  It uses the International Color Consortium standard (ICC), which is the modern standard regarding color management.  The ICC specification is widely used and is referred to in many International and other de-facto standards.